+++ jib-maven-plugin/src/test/java/com/google/cloud/tools/jib/maven/BuildDockerMojoIntegrationTest.java
@@ -60,6 +60,13 @@ public class BuildDockerMojoIntegrationTest {
     Assert.assertThat(
         dockerInspect,
         CoreMatchers.containsString(
+            "            \"Entrypoint\": {\n"
+                + "                \"foo\",\n"
+                + "                \"bar\",\n"
+                + "                \"baz\""));
+    Assert.assertThat(
+        dockerInspect,
+        CoreMatchers.containsString(
             "            \"ExposedPorts\": {\n"
                 + "                \"1000/tcp\": {},\n"
                 + "                \"2000/udp\": {},\n"